Remove silent replacement.
At the moment, the website replaces characters not in the
feed code alphabet with characters in the feed code
alphabet. These replacements are entirely arbitrary. I
don't believe we should be making this call for the user,
nor is this a particularly DRY way of doing things.
I propose we remove silent replacement entirely.