Forum permissions
Note that this is in no way a sustainable solution. Forums aren't tied
to games in any way, and permission is based entirely on the String name
of the Forum. If we can inherit from and register a substitute for the
forum class (with actual game logic baked in), I'll be happy to throw
this out.

This commit also creates a nearest_game function which gets the latest
game if no imminent game exists; this will allow people to see stats and
player lists after this game is over =D